fix: use double not float for mongo

+++ src/main/java/fr/insee/genesis/domain/service/editedprevious/EditedPreviousResponseJsonService.java
@@ -138,7 +138,7 @@ public class EditedPreviousResponseJsonService implements EditedPreviousResponse
                 return jsonParser.getIntValue();
             }
             case VALUE_NUMBER_FLOAT -> {
-                return jsonParser.getFloatValue();
+                return jsonParser.getDoubleValue();
             }
             case VALUE_TRUE, VALUE_FALSE -> {
                 return jsonParser.getBooleanValue();
